auth-service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
bcrypt==4.0.1
cryptography
prometheus-client
